news_storyboard/services/requirements.txt
newspaper3k
python-dotenv
openai
